Separate the client API from the CLI
This is a pre-work for issue #301
