+++ Makefile
@@ -6,6 +6,7 @@ API_ENDPOINT?=http://$(HOST):8080
 
 mkfile_path := $(abspath $(lastword $(MAKEFILE_LIST)))
 mkfile_dir := $(dir $(mkfile_path))
+WANAKU_CLI_CMD:= java -jar $(mkfile_dir)/cli/target/quarkus-app/quarkus-run.jar
 
 help:
 	@echo Make sure to adjust your environment to use Graal. For instance, run "sdk use ${GRAAL_VERSION}"
@@ -27,38 +28,38 @@ install:
 	ln -sf $(HOME)/bin/wanaku $(HOME)/bin/wk
 
 load-meta:
-	wanaku tools add -n "wanaku-tools-list" --description "List tools available on the Wanaku MCP router" --uri "http://localhost:8080/api/v1/tools/list" --type http
-	wanaku tools add -n "wanaku-resources-list" --description "List resources available on the Wanaku MCP router" --uri "http://localhost:8080/api/v1/resources/list" --type http
+	$(WANAKU_CLI_CMD) tools add -n "wanaku-tools-list" --description "List tools available on the Wanaku MCP router" --uri "http://localhost:8080/api/v1/tools/list" --type http
+	$(WANAKU_CLI_CMD) tools add -n "wanaku-resources-list" --description "List resources available on the Wanaku MCP router" --uri "http://localhost:8080/api/v1/resources/list" --type http
 
 test-resources:
-	wanaku resources expose --host $(API_ENDPOINT) --location=$(mkfile_dir)/tests/data/files/test.txt --mimeType=text/plain --description="Sample resource added via CLI" --name="sample-file" --type=file
-	wanaku resources expose --host $(API_ENDPOINT) --location=$(mkfile_dir)./tests/data/files --mimeType=text/plain --description="Sample resource dir added via CLI" --name="sample-dir" --type=file
+	$(WANAKU_CLI_CMD) resources expose --host $(API_ENDPOINT) --location=$(mkfile_dir)/tests/data/files/test.txt --mimeType=text/plain --description="Sample resource added via CLI" --name="sample-file" --type=file
+	$(WANAKU_CLI_CMD) resources expose --host $(API_ENDPOINT) --location=$(mkfile_dir)./tests/data/files --mimeType=text/plain --description="Sample resource dir added via CLI" --name="sample-dir" --type=file
 
 test-tools:
-	wanaku tools add --host $(API_ENDPOINT) -n "meow-facts" --description "Retrieve random facts about cats" --uri "https://meowfacts.herokuapp.com?count={count or 1}" --type http --property "count:int,The count of facts to retrieve" --required count
-	wanaku tools add --host $(API_ENDPOINT) -n "dog-facts" --description "Retrieve random facts about dogs" --uri "https://dogapi.dog/api/v2/facts?limit={count or 1}" --type http  --property "count:int,The count of facts to retrieve" --required count
-	wanaku tools add --host $(API_ENDPOINT) -n "camel-rider-quote-generator" --description "Generate a random quote from a Camel rider" --uri "file://$(mkfile_dir)/tests/data/routes/camel-route/hello-quote.camel.yaml" --type camel-yaml
-	wanaku tools add --host $(API_ENDPOINT) -n "tavily-search" --description "Search on the internet using Tavily" --uri "tavily://search" --type tavily
-	wanaku tools add --host $(API_ENDPOINT) -n "laptop-order" --description "Use the request system to order a new laptop" --uri "$(HOME)/.jbang/bin/camel run --max-messages=1 $(mkfile_dir)/tests/data/routes/camel-route/camel-jbang-quote.camel.yaml" --type exec
+	$(WANAKU_CLI_CMD) tools add --host $(API_ENDPOINT) -n "meow-facts" --description "Retrieve random facts about cats" --uri "https://meowfacts.herokuapp.com?count={count or 1}" --type http --property "count:int,The count of facts to retrieve" --required count
+	$(WANAKU_CLI_CMD) tools add --host $(API_ENDPOINT) -n "dog-facts" --description "Retrieve random facts about dogs" --uri "https://dogapi.dog/api/v2/facts?limit={count or 1}" --type http  --property "count:int,The count of facts to retrieve" --required count
+	$(WANAKU_CLI_CMD) tools add --host $(API_ENDPOINT) -n "camel-rider-quote-generator" --description "Generate a random quote from a Camel rider" --uri "file://$(mkfile_dir)/tests/data/routes/camel-route/hello-quote.camel.yaml" --type camel-yaml
+	$(WANAKU_CLI_CMD) tools add --host $(API_ENDPOINT) -n "tavily-search" --description "Search on the internet using Tavily" --uri "tavily://search" --type tavily
+	$(WANAKU_CLI_CMD) tools add --host $(API_ENDPOINT) -n "laptop-order" --description "Use the request system to order a new laptop" --uri "$(HOME)/.jbang/bin/camel run --max-messages=1 $(mkfile_dir)/tests/data/routes/camel-route/camel-jbang-quote.camel.yaml" --type exec
 
 clean-test-tools:
-	wanaku tools remove --name "meow-facts"
-	wanaku tools remove --name "dog-facts"
-	wanaku tools remove --name "camel-rider-quote-generator"
-	wanaku tools remove --name "tavily-search"
-	wanaku tools remove --name "laptop-order"
+	$(WANAKU_CLI_CMD) tools remove --name "meow-facts"
+	$(WANAKU_CLI_CMD) tools remove --name "dog-facts"
+	$(WANAKU_CLI_CMD) tools remove --name "camel-rider-quote-generator"
+	$(WANAKU_CLI_CMD) tools remove --name "tavily-search"
+	$(WANAKU_CLI_CMD) tools remove --name "laptop-order"
 
 clean-test-resources:
-	wanaku resources remove --name "sample-file"
-	wanaku resources remove --name "sample-dir"
+	$(WANAKU_CLI_CMD) resources remove --name "sample-file"
+	$(WANAKU_CLI_CMD) resources remove --name "sample-dir"
 
 clean-data: clean-test-resources clean-test-tools
-	wanaku resources list
-	wanaku tools list
+	$(WANAKU_CLI_CMD) resources list
+	$(WANAKU_CLI_CMD) tools list
 
 load-test: test-resources test-tools
-	wanaku targets resources list
-	wanaku targets tools list
+	$(WANAKU_CLI_CMD) targets resources list
+	$(WANAKU_CLI_CMD) targets tools list
 
 refresh-early-builds:
 	mvn -Pdist clean package
