+++ cli/src/main/java/ai/wanaku/cli/main/commands/forwards/ForwardsList.java
@@ -6,7 +6,7 @@ import jakarta.ws.rs.core.Response;
 import ai.wanaku.api.types.ForwardReference;
 import ai.wanaku.api.types.WanakuResponse;
 import ai.wanaku.cli.main.commands.BaseCommand;
-import ai.wanaku.cli.main.services.ForwardsService;
+import ai.wanaku.core.services.api.ForwardsService;
 import ai.wanaku.cli.main.support.PrettyPrinter;
 import io.quarkus.rest.client.reactive.QuarkusRestClientBuilder;
 import java.net.URI;
