+++ cli/src/main/java/ai/wanaku/cli/main/commands/resources/ResourcesRemove.java
@@ -7,7 +7,7 @@ import java.net.URI;
 
 import io.quarkus.rest.client.reactive.QuarkusRestClientBuilder;
 import ai.wanaku.cli.main.commands.BaseCommand;
-import ai.wanaku.cli.main.services.ResourcesService;
+import ai.wanaku.core.services.api.ResourcesService;
 import picocli.CommandLine;
 
 import static ai.wanaku.cli.main.support.ResponseHelper.commonResponseErrorHandler;
