+++ core/core-capabilities-base/src/main/resources/application.properties
@@ -1,6 +1,6 @@
 quarkus.qute.strict-rendering=false
 
 quarkus.log.level=WARNING
-quarkus.log.category."ai.wanaku".level=INFO
-%dev.quarkus.log.category."ai.wanaku".level=DEBUG
-%test.quarkus.log.category."ai.wanaku".level=INFO
+quarkus.log.category."ai.wanaku.core.capabilities".level=INFO
+%dev.quarkus.log.category."ai.wanaku.core.capabilities".level=DEBUG
+%test.quarkus.log.category."ai.wanaku.core.capabilities".level=INFO
