+++ core/core-services-api/src/main/java/ai/wanaku/core/services/api/ForwardsService.java
@@ -1,4 +1,4 @@
-package ai.wanaku.cli.main.services;
+package ai.wanaku.core.services.api;
 
 import jakarta.ws.rs.GET;
 import jakarta.ws.rs.POST;
