+++ core/core-services-api/src/main/java/ai/wanaku/core/services/api/TargetsService.java
@@ -1,4 +1,4 @@
-package ai.wanaku.cli.main.services;
+package ai.wanaku.core.services.api;
 
 import jakarta.ws.rs.Consumes;
 import jakarta.ws.rs.GET;
